worker/requirements.txt
numpy
Pillow
aiohttp
websockets
